--- apps/admin/src/features/Curriculum/Lesson/Components/Editor/fields/BlocksEditor.tsx
@@ -1,2 +1,0 @@
-// Re-exported from ExerciseEditor/BlocksEditor.tsx — see fields/index.ts
-export { BlocksEditor } from "../ExerciseEditor/BlocksEditor";
