--- apps/admin/src/features/Curriculum/Lesson/Components/Editor/fields/index.ts
@@ -1,2 +1,0 @@
-export { BlocksEditor } from "../ExerciseEditor/BlocksEditor";
-export { InteractionEditor } from "./InteractionEditor";
