+++ apps/admin/src/features/Lesson/detail/editor/templates.ts
