+++ apps/admin/src/features/Lesson/detail/preview/ExerciseInteraction.tsx
